add an `allSystems` list
this should be the only other list necessary. if tooling like
flake-utils is going to vendor this list, it should be centrally defined
and overridable when necessary without forking flake-utils.

The presence of `systems + "/all.nix"` requires that a downstream
`./flake.systems.nix` become a `./flake.systems` with the two files
`default.nix` and `all.nix`. `all.nix` can be copied from upstream and
updated whenever relevant. Any further downstreams won't notice if
`all.nix` is outdated because they'll be overriding it anyways with
either full upstream or their own file.